Begin smart screen rework
